backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
requests==2.32.3
pydantic==2.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
